refactor: swap imports to es module imports
