+++ test/sentenceTemplates.test.ts
@@ -1,7 +1,7 @@
-const { schema: schemaDefn } = require('../src');
-const { generateStatementSentence } = require('../src/sentenceTemplates');
+import { schema as schemaDefn } from '../src';
+import { generateStatementSentence } from '../src/sentenceTemplates';
 
-const examples = require('./testData/statementExamples.json');
+import examples from './testData/statementExamples.json';
 
 describe('generateStatementSentence', () => {
     describe('prognostic', () => {
@@ -243,7 +243,6 @@ describe('generateStatementSentence', () => {
             const { content } = generateStatementSentence(schemaDefn, examples[key]);
             expect(content.replace(' ({evidence})', '')).toEqual(result);
         });
-
 
         test('variant opposes diagnosis of disease', () => {
             const key = 'subject:Disease|conditions:Disease;PositionalVariant|relevance:opposes diagnosis';
